implement refund functions
